+++ order/src/main/java/org/liquidice/exmall/order/controller/OrderController.java
@@ -64,8 +64,8 @@ public class OrderController {
      * 计算订单总价
      */
     @GetMapping("/api/exmall/order/v1/order/total-price/{orderId}")
-    public Result<Long> calculateTotalPrice(@PathVariable("orderId") Long orderId) {
-        Long totalPrice = orderService.calculateTotalPrice(orderId);
+    public Result<Double> calculateTotalPrice(@PathVariable("orderId") Long orderId) {
+        Double totalPrice = orderService.calculateTotalPrice(orderId);
         return Results.success(totalPrice);
     }
 }
